requirements.txt
mwxml
mwparserfromhell
cassandra-driver
tqdm